Read player level as unsigned byte
v83 sends level as an unsigned byte (cap 200) but InPacket::read_byte
returns int8_t. For level 128+ (e.g. Stronkfeller at 160 -> 0xA0), the
signed value -96 widened into the uint16_t stats array as 65440, then
narrowed back to int16_t in Attack::playerlevel as -96. That inverted
the mob leveldelta: for a level-100 mob, leveldelta became 196 instead
of the intended 0-ish, which clamps min damage to 1 and hit chance to
~1%. Cast to uint8_t before widening at the three level-bearing call
sites (char load, stat update, incoming remote-attack level). Fixes
high-level characters seeing wrong level in UI and missing/hitting 1
on higher mobs on the WASM client.

+++ src/client/Net/Handlers/AttackHandlers.cpp
@@ -41,7 +41,10 @@ namespace jrc
         attack.type = type;
         attack.attacker = cid;
 
-        attack.level = recv.read_byte();
+        // Level is an unsigned byte on the wire (cap 200). read_byte returns
+        // int8_t; cast before widening so 128+ doesn't wrap into a negative
+        // playerlevel when damage math uses it.
+        attack.level = static_cast<uint8_t>(recv.read_byte());
         attack.skill = (attack.level > 0) ? recv.read_int() : 0;
 
         attack.display = recv.read_byte();
